# Use hex.EncodeToString to encode to hex